+++ cmd/samples/pageflow/src/server/errors/NameExistsError.js
@@ -1,7 +1,7 @@
 class NameExistsError extends Error {
   constructor(name) {
     super();
-    this.code = '400';
+    this.status = 400;
     this.message = `Product name '${name}' already exists.`
     this.name = 'NameExistsError';
   }
