+++ cmd/samples/pageflow/src/server/errors/NotFoundError.js
@@ -1,7 +1,7 @@
 class NotFoundError extends Error {
   constructor(id) {
     super();
-    this.code = '404';
+    this.status = 404;
     this.message = `Could not find product with id '${id}'.`;
     this.name = 'NotFoundError';
   }
