+++ cmd/samples/pageflow/src/server/errors/UnexpectedStatusError.js
@@ -1,7 +1,7 @@
 class UnexpectedStatusError extends Error {
   constructor(status) {
     super();
-    this.code = '400';
+    this.status = 400;
     this.message = `Unexpected status '${status}'.`
     this.name = 'UnexpectedStatusError';
   }
