+++ cmd/samples/pageflow/src/server/helpers/parseContent.js
@@ -1,5 +1,5 @@
 const parseContent = (str) => {
-  const parsedString = str.replace(/^"/, '').replace(/"$/, '').replace(/\\/g, '');
+  const parsedString = str.match(/^"(.*)"$/)[1];
   return JSON.parse(parsedString);
 };
