+++ cmd/samples/pageflow/src/server/middleware.js
@@ -1,0 +1,7 @@
+import cors from '@koa/cors';
+
+const initMiddleware = (app) => {
+  app.use(cors());
+};
+
+export default initMiddleware;
