+++ cmd/samples/pageflow/src/service/productService.js
@@ -1,6 +1,6 @@
 import config from '../server/config';
 
-const baseApiUrl = `${config.server.protocol}://${config.server.hostname}:${config.server.port}`;
+const baseApiUrl = `${config.server.protocol}://${config.server.hostname}:${config.server.port}/api`;
 
 const handleResponse = async (response) => {
   const data = await response.json();
